+++ asdf_astropy/converters/wcs/__init__.py
@@ -1,0 +1,6 @@
+__all__ = [
+    "WCSConverter",
+    "SlicedWCSConverter",
+]
+from .slicedwcs import SlicedWCSConverter
+from .wcs import WCSConverter