+++ CHANGES.rst
@@ -12,6 +12,8 @@
 
 - Add support for astropy.nddata.uncertainty classes [#239]
 
+- Add support for astropy.wcs.WCS and astropy.wcs.wcsapi.SlicedLowLevelWCS [#246]
+
 0.6.1 (2024-04-05)
 ------------------